Refactor: calculator 에서 자료형과 무관한 부분 로직 분리
