+++ src/main/java/Main.java
@@ -1,9 +1,5 @@
-import calculator.Calculator;
-import calculator.Result;
-import calculator.SentenceParser;
-import integerCalculator.IntegerCalculator;
+import calculator.*;
 import integerCalculator.IntegerSentenceParser;
-import integerCalculator.IntegerSequentialCalculator;
 
 import java.util.Scanner;
 
@@ -13,13 +9,13 @@ public class Main {
         String input = sc.nextLine();
 
         SentenceParser<Integer> sentenceParser = new IntegerSentenceParser();
-        Calculator<Integer> sequentialCalculator = new IntegerSequentialCalculator(sentenceParser.parse(input));
-        Calculator<Integer> calculator = new IntegerCalculator(sentenceParser.parse(input));
+        Calculator<Integer> sequentialCalculator = new SequentialCalculator<>(sentenceParser.parse(input));
+        Calculator<Integer> arithmeticCalculator = new ArithmeticCalculator<>(sentenceParser.parse(input));
 
         Result<Integer> sequentialResult = new Result<>(sequentialCalculator.makeAnswer());
         sequentialResult.print();
 
-        Result<Integer> result = new Result<>(calculator.makeAnswer());
-        result.print();
+        Result<Integer> arithmeticResult = new Result<>(arithmeticCalculator.makeAnswer());
+        arithmeticResult.print();
     }
 }
