+++ src/main/java/calculator/SequentialCalculator.java
@@ -1,19 +1,14 @@
-package integerCalculator;
+package calculator;
 
-import calculator.Calculator;
-import calculator.Operand;
-import calculator.Operator;
-
-public class IntegerSequentialCalculator extends Calculator<Integer> {
-    public IntegerSequentialCalculator(Operand<Integer> startOperand) {
+public class SequentialCalculator<T> extends Calculator<T>{
+    public SequentialCalculator(Operand<T> startOperand) {
         super(startOperand);
     }
-
     @Override
-    public Integer makeAnswer() {
+    public T makeAnswer() {
         boolean isOperand = true;
-        Operand<Integer> currentOperand = startOperand;
-        Operator<Integer> currentOperator = null;
+        Operand<T> currentOperand = startOperand;
+        Operator<T> currentOperator = null;
         while(true){
             if(isOperand){
                 if(currentOperand.getLatter()==null) break;
@@ -21,8 +16,8 @@ public class IntegerSequentialCalculator extends Calculator<Integer> {
                 isOperand = !isOperand;
                 continue;
             }
-            Operand<Integer> left = currentOperator.getFormer();
-            Operand<Integer> right = currentOperator.getLatter();
+            Operand<T> left = currentOperator.getFormer();
+            Operand<T> right = currentOperator.getLatter();
             currentOperand = left.calculate(currentOperator, right);
             if(right.getLatter()!=null) right.getLatter().connectFormer(currentOperand);
             isOperand = !isOperand;
